api/requirements.txt
fastapi
uvicorn
psycopg2
python-jose[cryptography]
passlib[bcrypt]